fix: migrate state to zustand
