+++ src/components/App.tsx
@@ -1,7 +1,17 @@
 import { useColorMode, useTheme } from "@chakra-ui/react";
-import { LinksProvider } from "@contexts/links";
-import { setStoredColorMode, StorageKey } from "@lib/webextension";
+import {
+  setNextStoredLinkId,
+  setStoredColorMode,
+  setStoredLinksAndKeys,
+  StorageKey,
+} from "@lib/webextension";
 import { parseColorMode } from "@models/color-mode";
+import {
+  parseLinkKeys,
+  parseNextLinkId,
+  parseStoredLinks,
+} from "@models/link-state";
+import { useLinkStore } from "hooks/useLinkStore";
 import * as React from "react";
 import browser from "webextension-polyfill";
 import Page from "./Page";
@@ -21,9 +31,36 @@ export default function App() {
     initializeColorMode();
   }, [initialColorMode, setColorMode]);
 
-  return (
-    <LinksProvider>
-      <Page />
-    </LinksProvider>
-  );
+  const setLinks = useLinkStore((state) => state.setLinks);
+  React.useEffect(() => {
+    const initializeState = async () => {
+      const {
+        [StorageKey.NEXT_LINK_ID]: storedNextLinkId,
+        [StorageKey.LINK_STORAGE_KEYS]: storedLinkKeys,
+      } = await browser.storage.local.get([
+        StorageKey.LINK_STORAGE_KEYS,
+        StorageKey.NEXT_LINK_ID,
+      ]);
+      const nextLinkId = parseNextLinkId(storedNextLinkId);
+      const linkKeys = parseLinkKeys(storedLinkKeys);
+
+      if (nextLinkId && linkKeys?.length) {
+        const storedLinks = parseStoredLinks(
+          await browser.storage.local.get(linkKeys)
+        );
+        const links =
+          (storedLinks && linkKeys.map((key) => storedLinks[key])) ?? [];
+        if (links.length) {
+          setLinks({ nextLinkId, linkKeys, links });
+          return;
+        }
+      }
+      // Cleanse stored state.
+      setStoredLinksAndKeys([], []);
+      setNextStoredLinkId(0);
+    };
+    initializeState();
+  }, []);
+
+  return <Page />;
 }
